dss-rava/requirements.txt
pyzabbix>=0.7,<0.8
python-novaclient
python-neutronclient
pylab
scipy
